# server-go/docs/swagger.yaml
basePath: /api
definitions:
...
  models.APIKey:
    properties:
      created:
        type: string
      expires_at:
        type: string
      id:
        description: UUID
        type: string
      last_used:
        type: string
      name:
        description: Human label
        type: string
      prefix:
        description: First 8 chars of the random token, plaintext
        type: string
      revoked:
        type: boolean
      revoked_at:
        type: string
      user_guid:
        description: Owner
        type: string
    type: object
...
  models.CreateAPIKeyRequest:
    properties:
      expires_at:
        description: Optional — null = never expires
        type: string
      name:
        type: string
    required:
    - name
    type: object
  models.CreateAPIKeyResponse:
    properties:
      api_key:
        $ref: '#/definitions/models.APIKey'
      token:
        description: PLAINTEXT — shown once, never persisted
        type: string
    type: object
...
host: localhost:3001
info:
  contact:
    email: support@example.com
    name: Dashboard Team
  description: Dashboard system with AI-powered chart generation
  title: TRVE Dashboards API
  version: "1.0"
paths:
...
  /api/api-keys:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.APIKey'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List the calling user's API keys
      tags:
      - api-keys
    post:
      consumes:
      - application/json
      parameters:
      - description: Key parameters
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CreateAPIKeyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.CreateAPIKeyResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create an API key for the calling user
      tags:
      - api-keys
  /api/api-keys/{id}:
    delete:
      parameters:
      - description: API key ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Revoke an API key
      tags:
      - api-keys
  /api/api-keys/all:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.APIKey'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List every API key in the deployment (admin)
      tags:
      - api-keys
...
swagger: "2.0"